refactor: file refactor and clean up
