+++ src/components/product-modal/ProductVariations.tsx
@@ -3,7 +3,7 @@ import type { CatalogsProductVariation } from "@moltin/sdk";
 import { useRouter } from "next/router";
 import { useContext } from "react";
 import { useEffect, useState } from "react";
-import { OptionDict } from "../../lib/product-types";
+import { OptionDict } from "../../lib/types/product-types";
 import {
   createEmptyOptionDict,
   ProductModalContext,
@@ -13,11 +13,11 @@ import {
   getOptionsFromSkuId,
   getSkuIdFromOptions,
   mapOptionsToVariation,
-  MatrixObjectEntry,
-} from "../../services/helper";
+} from "../../lib/product-helper";
 import ProductVariationStandard, {
   UpdateOptionHandler,
 } from "../product/variations/ProductVariationStandard";
+import { MatrixObjectEntry } from "../../lib/types/matrix-object-entry";
 
 interface IProductVariations {
   variations: CatalogsProductVariation[];
