+++ src/lib/types/matrix-object-entry.ts
@@ -1,0 +1,5 @@
+export type MatrixValue = string;
+
+export interface MatrixObjectEntry {
+  [key: string]: MatrixObjectEntry | MatrixValue;
+}
